Refactor to move Scene related components into a class
Related to #31

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/abrie/nl5/issues/31?shareId=XXXX-XXXX-XXXX-XXXX).
